Adding new features to the app

+++ src/main/java/net/engineeringdigest/journalApp/Service/JournalEntryService.java
@@ -7,6 +7,7 @@ import net.engineeringdigest.journalApp.entity.User;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -22,7 +23,7 @@ public class JournalEntryService {
      @Autowired
      private UserService userService;
 
-
+     @Transactional
      public void saveEntry(JournalEntry journalEntry, String userName){
 
          try{
@@ -30,10 +31,11 @@ public class JournalEntryService {
              journalEntry.setDate(LocalDateTime.now());
              JournalEntry saved = journalEntryRepository.save(journalEntry);
              user.getJournalEntries().add(saved);
-             userService.saveEntry(user);
+             userService.saveUser(user);
 
          } catch (Exception e){
              log.error("exception", e);
+             throw new RuntimeException("An error occurred while saving the entry", e);
          }
      }
     public void saveEntry(JournalEntry journalEntry){
@@ -54,10 +56,25 @@ public class JournalEntryService {
          return  journalEntryRepository.findById(Id);
      }
 
-     public void deleteById(ObjectId Id, String userName){
-         User user = userService.findByUserName(userName);
-         user.getJournalEntries().removeIf(x -> x.getId().equals(Id));
-         userService.saveEntry(user);
-         journalEntryRepository.deleteById(Id);
+     public boolean deleteById(ObjectId Id, String userName){
+         boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed  = user.getJournalEntries().removeIf(x -> x.getId().equals(Id));
+            if(removed){
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(Id);
+            }
+
+        } catch (Exception e){
+         System.out.println(e);
+         throw new RuntimeException("An error occured while saving the entry", e);
+        }
+        return removed;
+
+//     public List<JournalEntry> findByUserName(String userName){
+//         User user = userService.findByUserName(userName);
+//         return user.getJournalEntries() ;
+//
      }
 }
